rename into generate_linkml_from_pydantic.py since we make it generic
