+++ README.md
@@ -6,7 +6,7 @@ Translate pydantic-based schemas from python modules to LinkML.
 To run the program, use:
 
 ```
-  python3 generate_linkml_from_aind.py --root_module_name="aind_data_schema.models" --output_file="aind.yml"
+  python3 generate_linkml_from_pydantic.py --root_module_name="aind_data_schema.models" --output_file="aind.yml"
 ```
 
 ### Limitations
