+++ tests/test_converters.py
@@ -1,4 +1,4 @@
-from generate_linkml_from_aind import main
+from generate_linkml_from_pydantic import main
 
 def test_smoke():
     """Check that I can run this at all"""
